Add files via upload
This is a semi-automated Java application with Java Swing as the front end and csv files as the
database. It has a login portal, which enables both the faculty and student to login. The database
already has the information about the attendance of each student , for every course, allotted
over months. The database could be done through csv file reading from already prepared excel
sheets. The sheet consisted of the number of hours a student was present including late, excused
and unexcused. Using this information , both the normal percentage and the compensated
percentage can also be also calculated. A well- defined warning instructor can be implemented
in this project for faculties to know who all lag attendance . Attendance report can also be
viewed by the faculties subject wise on a monthly basis for all students whereas students can
view their attendance and compensated attendance in the student report. Modification of
student details can also be performed by faculties through their login. This is a unique , very
handy and yet useful project for administration of student attendance.